refactor : 에러메세지 부분 출력 수정
