+++ src/main/java/com/hanghae7/alcoholcommunity/domain/party/controller/PartyParticipateController.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.hanghae7.alcoholcommunity.domain.common.ResponseDto;
 import com.hanghae7.alcoholcommunity.domain.common.security.UserDetailsImplement;
-import com.hanghae7.alcoholcommunity.domain.party.dto.ParticipateResponseDto;
 import com.hanghae7.alcoholcommunity.domain.party.service.PartyParticipateService;
 
 import lombok.RequiredArgsConstructor;
