+++ src/main/java/com/hanghae7/alcoholcommunity/domain/party/dto/response/PartyDetailResponseDto.java
@@ -1,4 +1,4 @@
-package com.hanghae7.alcoholcommunity.domain.party.dto;
+package com.hanghae7.alcoholcommunity.domain.party.dto.response;
 
 import java.util.List;
