+++ src/main/java/com/hanghae7/alcoholcommunity/domain/party/dto/response/PartyResponseDto.java
@@ -1,16 +1,15 @@
-package com.hanghae7.alcoholcommunity.domain.party.dto;
+package com.hanghae7.alcoholcommunity.domain.party.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.hanghae7.alcoholcommunity.domain.member.entity.Member;
+import com.hanghae7.alcoholcommunity.domain.party.dto.Info.MemberInfoDto;
 import com.hanghae7.alcoholcommunity.domain.party.entity.Party;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -30,12 +29,6 @@ public class PartyResponseDto {
 	private Long partyId;
 	private String title;
 	private String content;
-/*	private String hostImage;
-	private String hostName;*/
-	// @JsonFormat(shape = JsonFormat.Shape.STRING,pattern = "yyyy-MM-dd'T'HH:mm",timezone = "Asia/Seoul")
-	// private LocalDateTime startDate;
-	// @JsonFormat(shape = JsonFormat.Shape.STRING,pattern = "yyyy-MM-dd'T'HH:mm",timezone = "Asia/Seoul")
-	// private LocalDateTime endDate;
 	@JsonFormat(shape = JsonFormat.Shape.STRING,pattern = "yyyy-MM-dd'T'HH:mm",timezone = "Asia/Seoul")
 	private LocalDateTime partyDate;
 	private Double latitude;
@@ -45,23 +38,8 @@ public class PartyResponseDto {
 	private int currentCount;
 	private LocalDateTime createdAt;
 	private LocalDateTime modifiedAt;
-
 	private List<MemberInfoDto> memberInfo;
 
-	/*public PartyResponseDto(Party party, String hostImage, String hostname) {
-		this.partyId = party.getPartyId();
-		this.title = party.getTitle();
-		this.content = party.getContent();
-		this.hostImage = hostImage;
-		this.hostName = hostname;
-		this.partyDate = party.getPartyDate();
-		this.latitude = party.getLatitude();
-		this.longitude = party.getLongitude();
-		this.totalCount = party.getTotalCount();
-		this.currentCount = party.getCurrentCount();
-		this.createdAt = party.getCreatedAt();
-		this.modifiedAt = party.getModifiedAt();
-	}*/
 	public PartyResponseDto(Party party) {
 		this.partyId = party.getPartyId();
 		this.title = party.getTitle();
