+++ src/main/java/com/hanghae7/alcoholcommunity/domain/party/service/PartyService.java
@@ -1,8 +1,6 @@
 package com.hanghae7.alcoholcommunity.domain.party.service;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -15,10 +13,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.hanghae7.alcoholcommunity.domain.common.ResponseDto;
 import com.hanghae7.alcoholcommunity.domain.member.entity.Member;
 
-import com.hanghae7.alcoholcommunity.domain.party.dto.ParticipateInfoDto;
-import com.hanghae7.alcoholcommunity.domain.party.dto.PartyDetailResponseDto;
-import com.hanghae7.alcoholcommunity.domain.party.dto.PartyRequestDto;
-import com.hanghae7.alcoholcommunity.domain.party.dto.PartyResponseDto;
+import com.hanghae7.alcoholcommunity.domain.party.dto.request.PartyRequestDto;
+import com.hanghae7.alcoholcommunity.domain.party.dto.response.PartyResponseDto;
 import com.hanghae7.alcoholcommunity.domain.party.entity.Party;
 
 import com.hanghae7.alcoholcommunity.domain.party.entity.PartyParticipate;
@@ -90,6 +86,7 @@ public class PartyService {
 	// 모임 게시글 수정
 	@Transactional
 	public ResponseEntity<ResponseDto> updateParty(Long partyId, PartyRequestDto partyRequestDto, Member member) {
+
 		Party party = partyRepository.findById(partyId).orElseThrow(
 			() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다.")
 		);
@@ -109,22 +106,22 @@ public class PartyService {
 	@Transactional
 	public ResponseEntity<ResponseDto> deleteParty(Long partyId, Member member) {
 
-		Party party = partyRepository.findById(partyId).orElseThrow(
-			() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
-		Member hostMember = partyParticipateRepository.findByPartyIdAndHost(partyId).orElseThrow(
-			()-> new IllegalArgumentException("없는파티임")
-		);
-		System.out.println("***넌누구냐!");
-		System.out.println(hostMember.getMemberUniqueId());
+		try {
+			Party party = partyRepository.findById(partyId).orElseThrow(
+				() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
 
-		if(!hostMember.getMemberUniqueId().equals(member.getMemberUniqueId())) {
-			return new ResponseEntity<>(new ResponseDto(400, "해당 사용자가 아닙니다."), HttpStatus.BAD_REQUEST);
-		} else {
-			partyRepository.delete(party);
-			return new ResponseEntity<>(new ResponseDto(200, "모임을 삭제하였습니다."), HttpStatus.OK);
+			Member hostMember = partyParticipateRepository.findByPartyIdAndHost(partyId).orElseThrow(
+				()-> new IllegalArgumentException("없는파티임")
+			);
+			if(!hostMember.getMemberUniqueId().equals(member.getMemberUniqueId())) {
+				return new ResponseEntity<>(new ResponseDto(400, "해당 사용자가 아닙니다."), HttpStatus.BAD_REQUEST);
+			} else {
+				partyRepository.delete(party);
+				return new ResponseEntity<>(new ResponseDto(200, "모임을 삭제하였습니다."), HttpStatus.OK);
+			}
+		}
+		catch (IllegalArgumentException e){
+			return new ResponseEntity<>(new ResponseDto(400, "해당 게시글이 존재하지 않거나, 없는 모임입니다."), HttpStatus.BAD_REQUEST);
 		}
 	}
-
-
-
 }
